+++ scio-core/src/main/scala/com/spotify/scio/values/SCollection.scala
@@ -1416,92 +1416,7 @@ sealed trait SCollection[T] extends PCollectionWrapper[T] {
   // Resource operations
   // =======================================================================
 
-  def resourceCollectFn[R, U: Coder](resource: R, resourceType: ResourceType)(
-    pfn: PartialFunction[(R, T), U]
-  ): DoFn[T, U] =
-    new DoFnWithResource[T, U, R] {
-      def getResourceType: ResourceType = resourceType
-      def createResource: R = resource
-      val isDefined = ClosureCleaner.clean(pfn.isDefinedAt(_)) // defeat closure
-      val g = ClosureCleaner.clean(pfn)
-      @ProcessElement def processElement(c: DoFn[T, U]#ProcessContext): Unit =
-        if (isDefined(getResource, c.element())) {
-          c.output(g(getResource, c.element()))
-        }
-    }
 
-  def resourceMapFn[R, U: Coder](resource: R, resourceType: ResourceType)(
-    f: (R, T) => U
-  ): DoFn[T, U] =
-    new DoFnWithResource[T, U, R] {
-      def getResourceType: ResourceType = resourceType
-      def createResource: R = resource
-      val g = ClosureCleaner.clean(f)
-      @ProcessElement def processElement(c: DoFn[T, U]#ProcessContext): Unit =
-        c.output(g(getResource, c.element()))
-    }
-
-  def resourceFlatMapFn[R, U: Coder](resource: R, resourceType: ResourceType)(
-    f: (R, T) => TraversableOnce[U]
-  ): DoFn[T, U] =
-    new DoFnWithResource[T, U, R] {
-      def getResourceType: ResourceType = resourceType
-      def createResource: R = resource
-      val g = ClosureCleaner.clean(f)
-      @ProcessElement def processElement(c: DoFn[T, U]#ProcessContext): Unit = {
-        val i = g(getResource, c.element()).toIterator
-        while (i.hasNext) c.output(i.next())
-      }
-    }
-
-  def resourceFilterFn[R](resource: R, resourceType: ResourceType)(
-    f: (R, T) => Boolean
-  ): DoFn[T, T] =
-    new DoFnWithResource[T, T, R] {
-      def getResourceType: ResourceType = resourceType
-      def createResource: R = resource
-      val g = ClosureCleaner.clean(f)
-      @ProcessElement def processElement(c: DoFn[T, T]#ProcessContext): Unit =
-        if (g(getResource, c.element())) {
-          c.output(c.element())
-        }
-    }
-
-  /**
-   * Return a new [[SCollection]] by applying a function that also takes in a resource and
-   * `ResourceType` to all elements of this SCollection.
-   */
-  def mapWithResource[R, U: Coder](resource: => R, resourceType: ResourceType)(
-    fn: (R, T) => U
-  ): SCollection[U] =
-    self.parDo(resourceMapFn(resource, resourceType)(fn))
-
-  /**
-   * Filter the elements for which the given `PartialFunction` that also takes in a resource and
-   * `ResourceType` is defined, and then map.
-   */
-  def collectWithResource[R, U: Coder](resource: => R, resourceType: ResourceType)(
-    pfn: PartialFunction[(R, T), U]
-  ): SCollection[U] =
-    self.parDo(resourceCollectFn(resource, resourceType)(pfn))
-
-  /**
-   * Return a new [[SCollection]] by first applying a function that also takes in a resource and
-   * `ResourceType` to all elements of this SCollection, and then flattening the results.
-   */
-  def flatMapWithResource[R, U: Coder](resource: => R, resourceType: ResourceType)(
-    fn: (R, T) => TraversableOnce[U]
-  ): SCollection[U] =
-    self.parDo(resourceFlatMapFn(resource, resourceType)(fn))
-
-  /**
-   * Return a new [[SCollection]] containing only the elements that satisfy a predicate that
-   * takes in a resource and `ResourceType`
-   */
-  def filterWithResource[R](resource: => R, resourceType: ResourceType)(
-    fn: (R, T) => Boolean
-  )(implicit coder: Coder[T]): SCollection[T] =
-    self.parDo(resourceFilterFn(resource, resourceType)(fn))(coder)
 
   // =======================================================================
   // Write operations
