+++ scio-test/src/test/scala/com/spotify/scio/values/SCollectionTest.scala
@@ -813,62 +813,4 @@ class SCollectionTest extends PipelineSpec {
       coll should containInAnyOrder(Seq(Seq.empty[Int]))
     }
   }
-
-  it should "support filter with resource" in {
-    runWithContext { sc =>
-      val p =
-        sc.parallelize(Seq(1, 2, 3, 4, 5))
-          .filterWithResource(new Semaphore(10, true), ResourceType.PER_INSTANCE) { (r, v) =>
-            r.acquire()
-            r.release()
-            v % 2 == 0
-          }
-      p should containInAnyOrder(Seq(2, 4))
-    }
-  }
-
-  it should "support flat map with resource" in {
-    runWithContext { sc =>
-      val p = sc
-        .parallelize(Seq("a b c", "d e", "f"))
-        .flatMapWithResource(new Semaphore(10, true), ResourceType.PER_INSTANCE) { (r, v) =>
-          r.acquire()
-          r.release()
-          v.split(" ")
-        }
-      p should containInAnyOrder(Seq("a", "b", "c", "d", "e", "f"))
-    }
-  }
-
-  it should "support map with resource" in {
-    runWithContext { sc =>
-      val p = sc
-        .parallelize(Seq("1", "2", "3"))
-        .mapWithResource(new Semaphore(10, true), ResourceType.PER_INSTANCE) { (r, v) =>
-          r.acquire()
-          r.release()
-          v.toInt
-        }
-      p should containInAnyOrder(Seq(1, 2, 3))
-    }
-  }
-
-  it should "support collect with resource" in {
-    runWithContext { sc =>
-      val records = Seq(
-        ("test1", 1),
-        ("test2", 2),
-        ("test3", 3)
-      )
-      val p = sc
-        .parallelize(records)
-        .collectWithResource(new Semaphore(10, true), ResourceType.PER_INSTANCE) {
-          case (r, ("test2", v)) =>
-            r.acquire()
-            r.release()
-            2 * v
-        }
-      p should containSingleValue(4)
-    }
-  }
 }
